pyproject.toml
[project]
name = "magnety"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.97.1",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]
